# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
        "version": "1.0.3",
        "title": "fundoo",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:4000",
    "basePath": "/",
    "tags": [
        {
            "name": "user",
            "description": "user login "
        }
    ],
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    },
    "paths": {
        "/registration": {
            "post": {
                "tags": [
                    "registration"
                ],
                "summary": "Create user",
                "description": "Add user details for registration",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/registration"
                        }
                    }
                ],
               
                "responses": {
                    "201": {
                        "description": "successfully created note"
                    },
                    "422": {
                        "description": "unprocessable entity"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "user"
                ],
                "description": "Returns ",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/user"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "422": {
                        "description": " unprocessable entity"
                    }
                }
            }
        },
        "/notes": {
            "post": {
                "tags": [
                    "note"
                ],
                "summary": "Create note",
                "description": "Add title and description to create note",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/note"
                        }
                    }
                ],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successfully created note"
                    },
                    "422": {
                        "description": "unprocessable entity"
                    }
                }
            },
            "get": {
                "tags": [
                    "note"
                ],
                "summary": "get all note",
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successfully get all note"
                    },
                    "422": {
                        "description": "unprocessable entity"
                    }
                }
            }
          
        },
        "/notes/5e5cbf6d279fd73b0281d3fb": {
        "put": {
            "tags": [
                "update"
            ],
            "summary": "search update",
            "description": "update title and description in note",
            "parameters": [
                {
                    "name": "noteId",
                    "in": "body",
                    "required": true,
                    "type": "string",
                    "schema": {
                        "$ref": "#/definitions/note"
                    }
                }
            ],
            "security": [
                {
                    "api_key": []
                }
            ],
            "responses": {
                "201": {
                    "description": "successfully get all note"
                },
                "422": {
                    "description": "unprocessable entity"
                }
            }
        }
    },

        "/archive": {
            "get": {
                "tags": [
                    "archive"
                ],
                "summary": "get archive note",
                "description": "get all archive note",
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successfully get all archive note"
                    },
                    "422": {
                        "description": "unprocessable entity"
                    }
                }
            }
        },
        "/trash": {
            "get": {
                "tags": [
                    "trash"
                ],
                "summary": "get trash note",
                "description": "get all trash note",
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successfully get all trash note"
                    },
                    "422": {
                        "description": "unprocessable entity"
                    }
                }
            }
        },
        "/label": {
            "get": {
                "tags": [
                    "label"
                ],
                "summary": "get  label",
                "description": "get all label ",
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successfully get all label "
                    },
                    "422": {
                        "description": "unprocessable entity"
                    }
                }
            },
            "post": {
                "tags": [
                    "label"
                ],
                "summary": "create label",
                "description": "create label ",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/label"
                        }
                    }
                ],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successfully create label "
                    },
                    "422": {
                        "description": "unprocessable entity"
                    }
                }
            }
        },
       
        "/allReminder": {
            "get": {
                "tags": [
                    "allReminder"
                ],
                "summary": "get Reminder note",
                "description": "get all Reminder note",
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successfully get all Reminder note"
                    },
                    "422": {
                        "description": "unprocessable entity"
                    }
                }
            }
            
        },
        "/search/zzzzzzz": {
            "get": {
                "tags": [
                    "search"
                ],
                "summary": "search note",
                
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successfully created note"
                    },
                    "422": {
                        "description": "unprocessable entity"
                    }
                }
            }
        }
    },
    "definitions": {
        "registration": {
            "type": "object",
            "properties": {
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "emailid": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            },
            "example": {
                "firstName": "riya",
                "lastName": "shukala",
                "emailid": "abc123@gmail.com",
                "password": "*********"
            }
        },
        "user": {
            "type": "object",
            "properties": {
                "emailid": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            },
            "example": {
                "emailid": "xyz@gmail.com",
                "password": "xxxxxxxxx"
            }
        },
        "note": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            },
            "example": {
                "title": "xyz",
                "description": "xyz"
            }
        },
        "label": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string"
                } 
            },
            "example": {
                "label": "xyz"
            }
        }
    }
}